Separate the uninstall and install steps as subcommands (#44)
* Separate the uninstall and install steps as subcommands

- Refactored/Renamed several names for better understanding
- Used separate uninstall and install configurations, because the uninstall case may rely on another set of artifacts or packages
- Updated deployment-configuration.yaml
- Updated README

* Incorporate pull request review inputs, simplify deployment-configuration.yaml

- Removed uinstallationDef and installationDef and using only one deploymentDef
- Uninstallation can be handled by using simply the old version of a release package, so no need to have uninstallationDef

* Remove build-scan plugin, because it is not used currently

+++ src/main/kotlin/io/sledge/deployer/common/Retry.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withTimeout
 
-suspend fun <T> retry(retries: Long = 5, delay: Long = 1000, block: (Long) -> T): T {
+suspend fun <T> retry(retries: Int = 5, delay: Long = 1000, block: (Long) -> T): T {
     for (i in 1..retries) {
         try {
             return withTimeout(5000) {
